Move utilities to their own package
